// src/frontend/webapp/modilist/src/services/swagger/api/models/update-address.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpdateAddress
 */
export interface UpdateAddress {
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    zipCode?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    city?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    district?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAddress
     */
    fullAddress?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateAddress
     */
    isDefault?: boolean;
}